requirements.txt
aiohttp~=3.11.16
configparser~=7.2.0
future~=1.0.0

# Basic modules
gensim

# Additional Packages for export Functionality
pandas>=2.0.0

# Extra libraries are installed when needed using pipmaster

pipmaster~=0.5.4
pydantic~=2.11.3
python-dotenv~=1.1.0

setuptools~=65.5.0
tenacity~=9.1.2

# LLM packages
tiktoken~=0.9.0
xlsxwriter>=3.1.0

numpy~=1.26.4
ascii_colors~=0.5.2
requests~=2.32.3
pathlib~=1.0.1
networkx~=3.4.2
nest-asyncio~=1.6.0
ollama~=0.4.7
httpx~=0.28.1
fitz~=0.0.1.dev2
docx~=0.2.4
matplotlib~=3.10.1
pyvis~=0.3.2

chromadb~=1.0.5
llama-index-core~=0.12.31
aiofiles~=24.1.0
tqdm~=4.67.1